Comments and Logout

+++ Cipher/src/main/java/com/revature/models/User.java
@@ -72,7 +72,7 @@ public class User implements Serializable, IUser {
 	}
 
 	@Override
-	public long getUserid() {
+	public int getUserid() {
 		return userid;
 	}
